Move cclw card component to cclw folder
