+++ themes/cclw/components/Articles.tsx
@@ -1,8 +1,8 @@
 import { Fragment } from "react";
 
-import { Card } from "@components/card/Card";
 import { ExternalLink } from "@components/ExternalLink";
 import { LinkWithQuery } from "@components/LinkWithQuery";
+import { Card } from "@cclw/components/Card";
 
 type TArticle = {
   url: string;
